Improve the NullInputPadder.
NullInputPadder behaves poorly in the case of short reads that are in
middle of a file, as you might see with stdin.

This commit improves the behavior in such cases, attempting to drain the
reader to an EOF before padding.

Additionally, some tests and hardening is added to improve reliability.